refactor: auth내의 user도메인 클래스 패키지 이동
