+++ src/main/java/yapp/buddycon/app/auth/application/service/Login.java
@@ -3,7 +3,7 @@ package yapp.buddycon.app.auth.application.service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import yapp.buddycon.app.auth.domain.User;
+import yapp.buddycon.app.user.domain.User;
 
 @Component
 @RequiredArgsConstructor
