+++ src/main/java/yapp/buddycon/app/user/adapter/UserEntity.java
@@ -1,4 +1,4 @@
-package yapp.buddycon.app.auth.adapter;
+package yapp.buddycon.app.user.adapter;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
