+++ src/main/java/yapp/buddycon/app/user/adapter/infra/JpaUserQueryStorage.java
@@ -1,9 +1,9 @@
-package yapp.buddycon.app.auth.adapter.infra;
+package yapp.buddycon.app.user.adapter.infra;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import yapp.buddycon.app.auth.application.port.out.UserQueryStorage;
-import yapp.buddycon.app.auth.domain.User;
+import yapp.buddycon.app.user.application.port.out.UserQueryStorage;
+import yapp.buddycon.app.user.domain.User;
 
 import java.util.Optional;
