+++ src/main/java/yapp/buddycon/app/user/adapter/infra/JpaUserRepository.java
@@ -1,7 +1,7 @@
-package yapp.buddycon.app.auth.adapter.infra;
+package yapp.buddycon.app.user.adapter.infra;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import yapp.buddycon.app.auth.adapter.UserEntity;
+import yapp.buddycon.app.user.adapter.UserEntity;
 
 import java.util.Optional;
